First commit for setup-basic child theme
This child theme “setup-basic” will ultimately try to be the most basic setup needed for any child theme.

It is also an attempt to create a model-based system using tower and github to centralize access to child theme formulas.

The goal is to maintain a development model that can pull or draw data from other locations for updates. This allows the duplication of child theme files to be used for different environments but with only one main source code.